backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.31.0
websockets==12.0
python-dotenv==1.0.1
mysqlclient==2.2.4
pydantic==2.6.3 